refactor: 공통 컴포넌트, 공통 hooks 분리, 일부 라인 조정
