+++ pages/edit-menu/[id].tsx
@@ -1,8 +1,6 @@
-import React, { useState, useEffect, useCallback } from 'react'
+import React, { useState, useEffect } from 'react'
 import styled from '@emotion/styled'
-import TagContainer from '@components/TagContainer'
-import ImageUploader from '@components/ImageUploader'
-import Button from '@components/Button'
+import { TagContainer, ImageUploader, Button } from '@components/common'
 import { Option } from '@interfaces'
 import { useMenu } from '@hooks/queries/useMenu'
 import { useChangeMenu } from '@hooks/mutations/useChangeMenuMutation'
@@ -11,8 +9,6 @@ import { InputList } from '@components/create-menu/InputList'
 import { useRecoilValue } from 'recoil'
 import { currentUser } from '@recoil/currentUser'
 
-import Spinner from '@components/Spinner'
-import useRouterLoading from '@hooks/useRouterLoading'
 export interface InputListType {
   franchiseId: number
   title: string
@@ -23,16 +19,22 @@ export interface InputListType {
 }
 
 const EditMenu = () => {
-  // 필드 값
-
+  const [file, setFile] = useState<File | null>(null)
+  const [isFileChange, setIsFileChange] = useState(false)
+  const [franchiseId, setFranchiseId] = useState(1)
+  const [title, setTitle] = useState('')
+  const [originalTitle, setOriginalTitle] = useState('')
+  const [optionList, setOptionList] = useState<Option[]>([])
+  const [expectedPrice, setExpectedPrice] = useState(0)
+  const [isPriceButtonClicked, setIsPriceButtonClicked] = useState(false)
+  const [tasteIdList, setTasteIdList] = useState<number[]>([])
+  const [isSubmitted, setIsSubmitted] = useState(false)
   const router = useRouter()
-
-  const isRouterLoading = useRouterLoading()
-
   const { id } = router.query
   const user = useRecoilValue(currentUser)
   const { mutate, isLoading: changeMenuLoading } = useChangeMenu()
   const { data: menuData, isLoading: getMenuLoading } = useMenu(Number(id))
+
   useEffect(() => {
     if (menuData) {
       if (menuData.author.id !== user.id) {
@@ -49,18 +51,6 @@ const EditMenu = () => {
       setIsPriceButtonClicked(!menuData.expectedPrice)
     }
   }, [menuData])
-  const [file, setFile] = useState<File | null>(null)
-  const [isFileChange, setIsFileChange] = useState(false)
-
-  const [franchiseId, setFranchiseId] = useState(1)
-  const [title, setTitle] = useState('')
-  const [originalTitle, setOriginalTitle] = useState('')
-  const [optionList, setOptionList] = useState<Option[]>([])
-  const [expectedPrice, setExpectedPrice] = useState(0)
-  const [isPriceButtonClicked, setIsPriceButtonClicked] = useState(false)
-  const [tasteIdList, setTasteIdList] = useState<number[]>([])
-
-  const [isSubmitted, setIsSubmitted] = useState(false)
 
   const checkButtonDisabled = () => {
     return !(
@@ -102,7 +92,7 @@ const EditMenu = () => {
     setTasteIdList(tagIdList)
   }
 
-  const handleEditSubmit = (e: React.FormEvent<HTMLButtonElement>) => {
+  const handleEditSubmit = () => {
     setIsSubmitted(true)
     const data = {
       franchiseId,
@@ -126,11 +116,6 @@ const EditMenu = () => {
 
   return (
     <FlexContainer>
-      {isRouterLoading || changeMenuLoading || getMenuLoading ? (
-        <Spinner />
-      ) : (
-        ''
-      )}
       <ImageUploaderWrapper>
         <ImageUploader
           value={menuData?.image}
