+++ pages/login/index.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled'
-import Input from '@components/Input'
+import { Input } from '@components/common'
 import {
   INPUT_EMAIL,
   INPUT_PASSWORD,
@@ -14,10 +14,10 @@ import {
 } from '@constants/inputConstants'
 import React, { useState, useCallback } from 'react'
 import { BsEye } from 'react-icons/bs'
-import Button from '@components/Button'
+import Button from '@components/common/Button'
 import { useLoginMutation } from '@hooks/mutations/useLoginMutation'
 import Link from 'next/link'
-import InputMessage from '@components/InputMessage'
+import InputMessage from '@components/common/InputMessage'
 
 interface Errors {
   email: string
