+++ src/components/Layout/Header.tsx
@@ -1,9 +1,9 @@
+import { useState, useCallback, useEffect } from 'react'
 import styled from '@emotion/styled'
 import { useRouter } from 'next/router'
 import { IoChevronBackSharp } from 'react-icons/io5'
 import { FiLogIn, FiUser } from 'react-icons/fi'
 import Link from 'next/link'
-import { useState, useCallback, useEffect } from 'react'
 import { MYINFO_URL, LOGIN_URL, USER_URL, HOME_URL } from '@constants/pageUrl'
 import { getToken } from '@utils/cookie'
 import { SMALL_LOGO } from '@constants/image'
