+++ src/components/Layout/index.tsx
@@ -1,5 +1,5 @@
-import styled from '@emotion/styled'
 import { ReactNode, useState, useEffect } from 'react'
+import styled from '@emotion/styled'
 import Header from './Header'
 import Navigation from './Navigation'
 import { useRouter } from 'next/router'
