+++ src/components/common/BannerCard.tsx
