+++ src/components/create-menu/InputList.tsx
@@ -1,11 +1,13 @@
 import styled from '@emotion/styled'
-import InputMessage from '@components/InputMessage'
-import Input from '@components/Input'
-import Button from '@components/Button'
+import {
+  Input,
+  InputMessage,
+  Button,
+  FranchiseSelect
+} from '@components/common'
 import { TiDelete } from 'react-icons/ti'
 import { BsCheckLg } from 'react-icons/bs'
 import { Option } from '@interfaces'
-import FranchiseSelect from '@components/FranchiseSelect'
 
 import {
   MAX_OPTION,
