+++ src/components/detail/MenuDetail.tsx
@@ -1,11 +1,10 @@
 import React, { useState } from 'react'
 import styled from '@emotion/styled'
-import Modal from '@components/Modal'
-import { BiDotsVerticalRounded, BiTrash } from 'react-icons/bi'
 import { Menu } from '@interfaces'
+import { Modal, Avatar } from '@components/common'
+import { BiDotsVerticalRounded, BiTrash } from 'react-icons/bi'
 import { useDeleteMenuMutation } from '@hooks/mutations/useDeleteMenuMutation'
 import { useRouter } from 'next/router'
-import Avatar from '@components/Avatar'
 import { BsFillPencilFill } from 'react-icons/bs'
 import { IoMdHeart } from 'react-icons/io'
 import { usePostLikeMutation } from '@hooks/mutations/usePostLikeMutation'
