+++ src/types/index.ts
@@ -1,52 +1,10 @@
 import theme from '@constants/theme'
+import { Taste } from '@interfaces'
 declare module '@emotion/react' {
   export interface Theme {
     color: Color
     layout: Layout
   }
-}
-
-export interface Taste {
-  id: number
-  name: string
-  color: string
-}
-export interface Option {
-  name: string
-  description: string
-}
-
-export interface Menu {
-  id: number
-  franchise: Franchise
-  image: string
-  title: string
-  originalTitle: string
-  author: User
-  content: string
-  optionList: Option[]
-  expectedPrice: number
-  tasteList: Taste[]
-  comments: number
-  likes: number
-  createdAt: string
-  updatedAt: string
-}
-
-export interface User {
-  id: number
-  image: string
-  role: string | null
-  nickname: string
-  email: string
-  createdAt: string
-  updatedAt: string
-}
-
-export interface Franchise {
-  id: number
-  image: string
-  franchise: string
 }
 
 type Color = typeof theme.color
